requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.9
en-core-web-sm==2.0.0
fake-useragent==0.1.11
future==0.17.1
Google-Search-API==1.1.14
idna==2.8
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==4.5.2
murmurhash==1.0.2
nltk==3.4
numpy==1.16.2
plac==0.9.6
preshed==2.0.1
PyYAML==3.13
regex==2018.1.10
requests==2.21.0
selenium==2.53.6
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
spacy==2.0.18
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
ujson==1.35
Unidecode==1.0.23
urllib3==1.24.1
vcrpy==2.0.1
wrapt==1.10.11
yarl==1.3.0
